Add files via upload
his website is a modern, responsive portfolio and contact platform built to showcase my skills, projects, and personality. Designed with HTML, CSS, and JavaScript, it features an engaging user interface with smooth transitions and visual elements. The "Reach Me" section allows visitors to send messages directly to my email using EmailJS, with real-time input validation and a clean form layout.

The site also includes a footer with social media icons, making it easy to connect on platforms like LinkedIn and GitHub. Whether you're a recruiter, collaborator, or visitor, this site is crafted to be intuitive, informative, and visually appealing.